+++ src/dateCalculate/index.ts
@@ -7,49 +7,34 @@ import {formatDate} from "../index";
  * @param flag：对 年月日时分秒 那个进行计算
  * @param format：对 计算后的日期进行格式化的格式，默认使用 YY-MM-DD hh:mm:ss。比如：2020-01-02 08:00:00
  */
-declare let duration: object;
-
-
-class DateCalculate {
-  static dateCalculate = (dateStr: string | Date, num: number, flag: string, format?: string): string => {
-    let date = new Date(dateStr)
-    switch (flag) {
-      case 'year':
-        date.setFullYear(date.getFullYear() + num);
-        break
-      case 'month':
-        date.setMonth(date.getMonth() + num);
-        break
-      case 'day':
-        date.setDate(date.getDate() + num);
-        break
-      case 'hour':
-        date.setHours(date.getHours() + num);
-        break
-      case 'minute':
-        date.setMinutes(date.getMinutes() + num);
-        break
-      case 'second':
-        date.setSeconds(date.getSeconds() + num);
-        break
-      default:
-        throw new Error("时间计算错误")
-    }
-    return formatDate(date.getTime(), format)
+const dateCalculate = (dateStr: string | Date, num: number, flag: string, format?: string): string => {
+  let date = new Date(dateStr)
+  switch (flag) {
+    case 'year':
+      date.setFullYear(date.getFullYear() + num);
+      break
+    case 'month':
+      date.setMonth(date.getMonth() + num);
+      break
+    case 'day':
+      date.setDate(date.getDate() + num);
+      break
+    case 'hour':
+      date.setHours(date.getHours() + num);
+      break
+    case 'minute':
+      date.setMinutes(date.getMinutes() + num);
+      break
+    case 'second':
+      date.setSeconds(date.getSeconds() + num);
+      break
+    case 'milliseconds':
+      date.setMilliseconds(date.getMilliseconds() + num);
+      break
+    default:
+      throw new Error("时间计算有误，不支持 年月日时分秒毫秒 以外的计算")
   }
-
-  /**
-   * 计算两个时间的差
-   * @param start：开始时间
-   * @param end：结束时间
-   */
-  static timeSub = (start: string | Date, end: string | Date): number => {
-    let startDate = new Date(start)
-    let endDate = new Date(end)
-    let date = endDate.getTime() - startDate.getTime()
-    let misSecond = Math.floor(date);
-    return misSecond
-  }
+  return formatDate(date.getTime(), format)
 }
 
-export default DateCalculate;
+export default dateCalculate;
