+++ src/index.ts
@@ -57,7 +57,8 @@ export { default as getDomText } from './getDomText';
 export { default as getDomPageXY } from './getDomPageXY';
 export { default as getDomScreenXY } from './getDomScreenXY';
 export { default as xss } from './xss';
-export { default as DateCalculate } from './dateCalculate';
+export { default as dateCalculate } from './dateCalculate';
+export { default as timeSub } from './timeSub';
