+++ test/format.test.js
@@ -7,7 +7,8 @@ import {
   transformArray,
   url2obj,
   arrayToListNode,
-  DateCalculate,
+  dateCalculate,
+  timeSub,
 } from '../src/index';
 describe('数据结构相关测试', () => {
   test('数据深拷贝', () => {
@@ -245,11 +246,11 @@ describe('数据结构相关测试', () => {
   test('日期计算测试', () => {
     // 计算n天前或者n天后的时间
     let date = new Date('2020-01-01')
-    expect(DateCalculate.dateCalculate(date, -1, 'year')).toEqual('2019-01-01 08:00:00')
+    expect(dateCalculate(date, -1, 'year')).toEqual('2019-01-01 08:00:00')
 
 
     let start = new Date('2020-01-01 08:00:08')
     let end = new Date('2020-01-01 09:00:08')
-    expect(DateCalculate.timeSub(start, end)).toEqual(3600000)
+    expect(timeSub(start, end)).toEqual(3600000)
   });
 });
